Parameters: remove need for most getattr calls using caching
If a Parameter is marked as _inherited=True, then it does not store its default
value (for example) explicitly but rather inherits the value in real time
from its closest non-inherited parent. This was implemented by deletion of
the attribute and repeated getattr calls to find the correct inherited value.

This commit implements storage of a reference to the closest uninherited
parent, which is recalculated lazily in the uncommon event that the correct
parent is changed (by explicitly setting the default value for a previously
inherited parent, or by calling Parameter.reset() on an uninherited parent in
some cases).