SharedParameter: improve source detection
in the case that the values attribute of a Parameter is not yet
populated, the source of a SharedParameter would not be detectable. We
can instead find a corresponding source Parameter using the default.
This should only be used for completely stateless purposes (checking the
value of a "descriptive" Parameter attribute like "stateful" or
"loggable")