components: add check_user_specified decorator for _user_specified
Replaces previous determination of the _user_specified flag of a
Parameter simply by checking if the value received for a parameter in
Component.__init__ is None or not. Instead, a decorator is added to all
Component.__init__ methods to capture arguments passed in by the creator
of the Component (or additionally within the __init__ method of a
subclass).

This allows:
 - Parameter default values to be specified as defaults in __init__
 methods
 - with the specify_none flag, a Parameter value of None to be
 user-specified