Component: __init__: use fully parsed function if available
function is uniquely handled by _instantiate_function, which accepts a
function argument. This is generally what is passed in to the
constructor, but in some cases this template may be different than what
is stored as the parameter value from _initialize_parameters. This could
cause shared parameters to be missed if function is already
instantiated.