DriftDiffusionIntegrator: limit 3d->2d casting
When variable is 2d, value will be 3d. In the case that the variable is
just a single value in a 2d array, we want the array to remain 2d
(7d9b3294fc122702d49d84f82fee0a7c8c7a1c3e). This commit avoids changing
variable shapes when unnecessary for multi-element 2d arrays.