pytorchcomponents: add semi-deep __deepcopy__
Several PNL objects override __deepcopy__ to only copy
references (instead of full new objects) to certain other PNL components
in order for them to be used sensibly within the context of working
with a model in PNL (ex: Functions store references 'owner' to the
Components they belong to. Instead of copying owners as well, which
would result in ghost objects that don't belong to user models, only
references to the original owners are stored, and the other relevant
attributes are deepcopied)

This creates similar behavior for PytorchMechanismWrapper and
PytorchProjectionWrapper so that they can be copied across contexts
maintaining references to the original PNL objects rather creating new
ones that will not be used properly.